feat: Added in checkAuthority middleware, Changed registerUser and added LoginUser
Still need to implement checkAuthority. registerUser now creates a token and login user uses username to find a user and creates a token for that user. Still need to make changes to getAllUsers to remove sensitive information on call and add specific error handling for 404 errors to user operations.

BREAKING CHANGE